change 'steps_provided' from List to Set; improve test coverage
